requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.5.2
apturl==0.5.2
asgiref==3.8.1
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
click==8.1.8
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
debugpy==1.8.11
decorator==4.4.2
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
Django==2.1
dnspython==2.6.1
duplicity==0.8.12.0
email-validator==2.2.0
entrypoints==0.3
et-xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.6
fastapi-cli==0.0.7
fasteners==0.14.1
future==0.18.2
gpg==1.13.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.14.0
httptools==0.6.4
httpx==0.28.1
idna==2.8
importlib-metadata==8.5.0
importlib-resources==6.4.5
iniconfig==2.0.0
ipython-genutils==0.2.0
itsdangerous==2.2.0
jinja2==3.1.5
jupyter-core==4.6.3
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
monotonic==1.5
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
openpyxl==3.1.5
orjson==3.10.12
packaging==20.3
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
pluggy==1.5.0
protobuf==3.6.1
psycopg2-binary==2.9.10
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pydantic==2.10.4
pydantic-core==2.27.2
pydantic-extra-types==2.10.1
pydantic-settings==2.7.0
pygments==2.18.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyRFC3339==1.1
pytest==8.3.4
python-apt==2.0.0+ubuntu0.20.4.3
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
rich==13.9.4
rich-toolkit==0.12.0
SecretStorage==2.3.1
shellingham==1.5.4
simplejson==3.16.0
six==1.14.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.5.0
starlette==0.41.3
systemd-python==234
tomli==2.2.1
traitlets==4.3.3
typer==0.15.1
typing-extensions==4.12.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
ujson==5.10.0
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
uvicorn==0.33.0
uvloop==0.21.0
wadllib==1.3.3
watchfiles==0.24.0
websockets==13.1
xkit==0.0.0
zipp==3.20.2
